requirements.txt
Flask>=3.0.3
requests>=2.32.3
beautifulsoup4>=4.12.3
loguru>=0.7.0
python-dotenv>=1.0.1